Fix multiple issues relating to font rendering
When numberOfLines == 1, properly convert line breaks to spaces.
When converting newlines to spaces, render properly even when adjusting the font size.
Render the truncation character properly when truncating due to numberOfLines.
Render the truncation character properly when truncating after a blank line.

Also remove the ugly hack for guessing at glyphs when the font tables could not be read.
This hack only worked for some fonts, and it's better to fail than to possibly
render the wrong text.