Mark a wrap point before processing the glyph instead of after
When we encounter the need to wrap, we backtrack to the last wrap point and
insert a line break. If we find the need to wrap while processing a space,
we end up backtracking to the previous space instead. By inserting a wrap
point before checking the line length, we can wrap right at the current
space instead.