Missing category at runtime for static lib
There's a bug (misfeature?) in the iOS development kit that causes object files containing only categories to be skipped, when the file in question is built as a static library. This results in runtime errors, of course.

This quick fix defines an empty class, which is enough to force the linker to include the category when building the final executable.